Publish new version of package packages/static to npm package repo.

+++ src/packages/static/production-build.py
@@ -27,7 +27,8 @@ try:
     # This is necessary for now, since webpack gets content -- such as the primus script that hardcodes the base url --
     # from webapp-lib, so for our bundle files to be correct, we have to build webapp-lib.
     cmd('BASE_URL="" npm run build', '../../webapp-lib')
-    shutil.rmtree('dist')
+    if os.path.exists('dist'):
+        shutil.rmtree('dist')
     cmd("NODE_OPTIONS=--max_old_space_size=8000 COCALC_BASE_URL='/' BABEL_ENV=production NODE_ENV=production webpack --progress --color"
         )
 finally:
